assignment/inference/requirements.txt
fastapi==0.116.1
uvicorn==0.23.2
Pillow==10.1.0
tensorflow>=2.16.0
huggingface_hub>=0.20.0
python-multipart
